[refactor] Extracted extensions/dsl from extensions/view (4m)
